requirements.txt
fastapi==0.115.9
uvicorn==0.34.0
sqlalchemy==2.0.35
asyncpg==0.29.0
alembic==1.13.3
pydantic==2.10.6
pydantic-settings==2.5.2
passlib==1.7.4
PyJWT==2.10.1
python-multipart==0.0.20
bcrypt==4.0.0
